refactor(ai): extract shared chat/completions adapter from ollama driver
